requirements.txt
streamlit>=1.35.0
tavily-python>=0.4.0
python-dotenv>=1.0.1
openai>=1.55.3
google-generativeai>=0.7.2
